Add a master repair job using an external reparent
This change allows to repair a master when there appears to be none or existing master is unhealthy. The algorithm should converge to one master but there's a chance of more than one master for a short period of time. With random jittering for the check interval this should speed up the convergence.

Signed-off-by: Karel Alfonso Sague <kalfonso@squareup.com>